% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_keys.R
\name{add_keys}
\alias{add_keys}
\title{Add primary and foreign keys automatically}
\usage{
add_keys(dm)
}
\arguments{
\item{dm}{A \code{dm} object.}
}
\value{
An updated dm with primary and foreign key relations.
}
\description{
This function adds primary and foreign keys to a 'dm' object automatically,
assuming the names of the tables and keys follow these rules:
\enumerate{
\item For each table, the primary key has the same name as the table with the
prefix \verb{_id}. For example the table \code{x} has a primary key named \code{x_id}.
\item If an \verb{_id} column exists in a table which name is different than the
prefix of \verb{_id}, then that column is a foreign key linking to another table.
For example, in the table \code{x_id}, the column \code{y_id} is a foreign key linking
to the table \code{y}.
}
}
\examples{
dm <- dm(
  x = data.frame(x_id = 1, y_id = 2),
  y = data.frame(y_id = 2)
)

add_keys(dm)
}
