% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm-package.R
\docType{package}
\name{adm-package}
\alias{adm-package}
\alias{_PACKAGE}
\alias{testduplicate-package}
\title{adm: Automatic Data Models}
\description{
Create data models, adding primary and foreign keys automatically from the names of tables and key columns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/adm}
  \item Report bugs at \url{https://github.com/maurolepore/adm/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]

}
\keyword{internal}
