% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalADM.R
\name{adm}
\alias{adm}
\title{Multidimensional Embedding Visualization (ADM)}
\usage{
adm(
  e,
  k.dim = NULL,
  dist.power = 0.5,
  conn.prop = 0.003,
  raw.d.pwr = 0.5,
  diffu.steps = NA,
  diffu.factor = 3.5,
  distr.template = "combine",
  gamma.shape = 3,
  gamma.rate = 3,
  scale.dist = TRUE,
  symmetrize = "mean",
  dist.quantile = 0.25
)
}
\arguments{
\item{e}{A list of matrices, each representing a dimension reduction result.}

\item{k.dim}{Number of dimensions to use. If NULL, uses the number of columns in the first matrix of e.}

\item{dist.power}{Power to raise distances to when merging. Default is 0.5.}

\item{conn.prop}{Proportion of connections to consider. Default is 0.02.}

\item{raw.d.pwr}{Power for raw distance calculation. Default is 0.5.}

\item{diffu.steps}{Steps for diffusion. If NA, calculated based on graph properties.}

\item{diffu.factor}{Factor to multiply diffusion steps by. Default is 3.5.}

\item{distr.template}{Distribution template to use. Options are "none", "combine", "gamma", or "parametric". Default is "gamma".}

\item{gamma.shape}{Shape parameter for gamma distribution. Default is 3.}

\item{gamma.rate}{Rate parameter for gamma distribution. Default is 3.}

\item{scale.dist}{Whether to scale distances. Default is TRUE.}

\item{symmetrize}{Method to symmetrize distance matrix. Options are "mean", "min", or "max". Default is "mean".}

\item{dist.quantile}{Quantile to use for distance calculations. Default is 0.25.}
}
\value{
A list containing the final merged distance matrix.
}
\description{
Multidimensional Embedding Visualization (ADM)
}
