% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_eval.R
\name{adm_eval}
\alias{adm_eval}
\title{Calculate different model performance metrics}
\usage{
adm_eval(obs, pred)
}
\arguments{
\item{obs}{numeric. Observed abundance}

\item{pred}{umeric. Predicted abundance}
}
\value{
a tibble with next columns
}
\description{
Calculate different model performance metrics
}
