% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_extract.R
\name{adm_extract}
\alias{adm_extract}
\title{Extract environmental data values from a spatial raster based on x and y coordinates}
\usage{
adm_extract(data, x, y, env_layer, variables = NULL, filter_na = TRUE)
}
\arguments{
\item{data}{data.frame. Database with species abundance with x and y coordinates}

\item{x}{character. Column name with spatial x coordinates}

\item{y}{character. Column name with spatial y coordinates}

\item{env_layer}{SpatRaster. Raster or raster stack with environmental variables.}

\item{variables}{character. Vector with the variable names of predictor (environmental) variables
Usage variables. = c("elevation", "sand", "cfvo"). If no variable is specified, function will
return data for all layers. Default NULL}

\item{filter_na}{logical. If filter_na = TRUE (default), the rows with NA values for any of the
environmental variables are removed from the returned tibble.}
}
\value{
A tibble that returns the original data base with additional columns
for the extracted environmental variables at each x y location from the SpatRaster object used
in 'env_layer'
}
\description{
Extract environmental data values from a spatial raster based on x and y coordinates
}
\examples{
\dontrun{
require(terra)

# Load datasets
data("sppabund")
envar <- system.file("external/envar.tif", package = "adm")
envar <- terra::rast(envar)

# Extract data
some_sp <- sppabund \%>\%
  filter(species == "Species one") \%>\% dplyr::select(species, ind_ha, x,y)

# Extract environmental data from envar raster for all locations in spp
ex_spp <-
  adm_extract(
    data = some_sp,
    x = "x",
    y = "y",
    env_layer = envar,
    variables = NULL,
    filter_na = FALSE
  )

# Extract environmental for two variables and remove rows with NAs
ex_spp2 <-
  adm_extract(
    data = some_sp,
    x = "x",
    y = "y",
    env_layer = envar,
    variables = c("bio1", "elevation"),
    filter_na = TRUE
  )

ex_spp
ex_spp2

}
}
