% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cretusa_data}
\alias{cretusa_data}
\title{A data set containing abundance of Cynophalla retusa.}
\format{
A tibble with 366 rows and 6 variables:
\describe{
  \item{species}{species names}
  \item{ind_ha}{species abundance expressed as individuals per hectare}
  \item{x}{longitude of species occurrences}
  \item{y}{latitude of species occurrences}
  \item{.part}{partition folds}
}
}
\usage{
cretusa_data
}
\description{
Cynophalla retusa (Griseb.) Cornejo & Iltis (Capparaceae) is a 
shrub native to northeastern Argentina, Paraguay, Bolivia, 
and central Brazil.
}
\examples{
\dontrun{
require(dplyr)
data("cretusa_data")
cretusa_data
}
}
\keyword{datasets}
