% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cretusa_predictors}
\alias{cretusa_predictors}
\title{Raster with Principal Component}
\format{
A raster in tif format with the first principal components.
}
\usage{
cretusa_predictors()
}
\description{
A raster with principal components derived by a principal component analysis based
on climatic (from Chelsa: chelsa-climate.org) and edaphic (from SoilGrids: soilgrids.org) variables.
}
\examples{
\dontrun{
require(terra)
envar <- system.file("external/cretusa_predictors.tif", package = "adm")
envar <- terra::rast(envar)
plot(envar)
}
}
