% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{croppin_hood}
\alias{croppin_hood}
\title{Crop rasters to build samples for Convolutional Neural Networks}
\usage{
croppin_hood(occ, x, y, raster, size)
}
\arguments{
\item{occ}{}

\item{x}{}

\item{y}{}

\item{raster}{}

\item{size}{}
}
\description{
Crop rasters to build samples for Convolutional Neural Networks
}
