% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_abund_bpdp.R
\name{data_abund_bpdp}
\alias{data_abund_bpdp}
\title{Calculate data to construct partial dependence surface plots}
\usage{
data_abund_bpdp(
  model,
  predictors,
  resolution = 50,
  training_data = NULL,
  training_boundaries = NULL,
  projection_data = NULL,
  clamping = FALSE
)
}
\arguments{
\item{model}{}

\item{predictors}{}

\item{resolution}{}

\item{training_data}{}

\item{training_boundaries}{}

\item{projection_data}{}

\item{clamping}{}
}
\description{
Calculate data to construct partial dependence surface plots
}
