% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model.R
\name{data_model}
\alias{data_model}
\title{Create a data model}
\usage{
data_model(dir)
}
\arguments{
\item{dir}{A directory with .csv tables.}
}
\value{
A "dm" object.
}
\description{
Create a data model
}
\examples{
library(tibble)
library(withr)
library(readr)
library(fs)

tmp <- local_tempdir()
write_csv(tibble(x_id = 1, y_id = 2), path(tmp, "x.csv"))
write_csv(tibble(y_id = 2), path(tmp, "y.csv"))

data_model(tmp)
}
