% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cal_functions.R
\name{dist.grp}
\alias{dist.grp}
\title{Calculate distance-based group statistics}
\usage{
dist.grp(distmat, grp, k = 1:20)
}
\arguments{
\item{distmat}{A distance matrix.}

\item{grp}{A vector of group labels for each data point.}

\item{k}{Vector of integers; the number of neighbors to consider.}
}
\value{
A matrix with two columns: the k values and corresponding distance-based group statistics.
}
\description{
This function calculates distance-based group statistics for a given distance matrix and grouping.
}
