% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{envar}
\alias{envar}
\title{Raster with environmental data}
\format{
A raster in tif format with the following variables.
\describe{
  \item{bio1}{Annual mean temperature}
  \item{bio12}{Annual precipitation}
  \item{bio15}{Precipitation seasonality}
  \item{bio3}{Isothermality}
  \item{cfvo}{Volumetric fraction of coarse fragments}
  \item{elevation}{Elevation}
  \item{sand}{Soil sand content}
  \item{eco}{Terrestrial ecoregion}
}
}
\usage{
envar()
}
\description{
A raster with climatic (from Chelsa: chelsa-climate.org), edaphic (from SoilGrids: soilgrids.org), and ecoregion data (from WWF https://www.worldwildlife.org/publications/terrestrial-ecoregions-of-the-world).
}
\examples{
\dontrun{
require(terra)
envar <- system.file("external/envar.tif", package = "adm")
envar <- terra::rast(envar)
plot(envar)
}
}
