% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{family_selector}
\alias{family_selector}
\title{Select probability distributions for GAM and GLM}
\usage{
family_selector(data, response)
}
\arguments{
\item{data}{data.frame or tibble. Database with species abundance}

\item{response}{character. Column name with species abundance}
}
\value{
tibble with family_name, family_call, range, and discrete columns (if family distribution
or not discrete)
}
\description{
Select probability distribution available in gamlss.dist suited for a given response variables (e.g., count, zero-inflated) used to fit GAM and GLM models. See \link[gamlss.dist]{gamlss.family} for more details.
}
\examples{
\dontrun{
data(sppabund)

family_selector(data = sppabund, response = "ind_ha")
}

}
