% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_abund_raf.R
\name{fit_abund_raf}
\alias{fit_abund_raf}
\title{Fit and validate Random Forests models}
\usage{
fit_abund_raf(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition = NULL,
  hold_out_set = NULL,
  predict_part = FALSE,
  mtry = length(c(predictors, predictors_f))/3,
  ntree = 500,
  verbose = TRUE
)
}
\arguments{
\item{data}{tibble or data.frame. Database with response, predictors, and partition values}

\item{response}{character. Column name with species abundance.}

\item{predictors}{character. Vector with the column names of quantitative predictor variables (i.e. continuous variables). Usage predictors = c("temp", "precipt", "sand")}

\item{predictors_f}{character. Vector with the column names of qualitative predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")}

\item{fit_formula}{formula. A formula object with response and predictor variables (e.g. formula(abund ~ temp + precipt + sand + landform)). Note that the variables used here must be consistent with those used in response, predictors, and predictors_f arguments. Default NULL}

\item{partition}{character. Column name with training and validation partition groups.
If partition = NULL, the model will be validated with the same data used for fitting.}

\item{hold_out_set}{tibble or data.frame. Optional hold-out dataset, never used during training, with the response and predictor columns. If provided, performance metrics are also calculated on it. Default NULL}

\item{predict_part}{logical. Save predicted abundance for testing data. Default is FALSE.}

\item{mtry}{numeric. Number of variables randomly sampled as candidates at each split. Default (length(c(predictors, predictors_f))/3)}

\item{ntree}{numeric. Number of trees to grow. This should not be set to too small a number, to ensure that every input row gets predicted at least a few times. Default 500}

\item{verbose}{logical. If FALSE, disables all console messages. Default TRUE}
}
\value{
A list object with:
\itemize{
\item model: A "randomForest" class object from randomForest package. This object can be used for predicting.
\item predictors: A tibble with quantitative (c column names) and qualitative (f column names) variables use for modeling.
\item performance: Averaged performance metrics (see \code{\link{adm_eval}}).
\item performance_part: Performance metrics for each replica and partition.
\item predicted_part: Observed and predicted abundance for each test partition.
}
}
\description{
Fit and validate Random Forests models
}
\examples{
\dontrun{
require(terra)
require(dplyr)

# Database with species abundance and x and y coordinates
data("sppabund")

# Extract data for a single species
some_sp <- sppabund \%>\%
  dplyr::filter(species == "Species one") \%>\%
  dplyr::select(-.part2, -.part3)

# Explore reponse variables
some_sp$ind_ha \%>\% range()
some_sp$ind_ha \%>\% hist()

# Here we balance number of absences
some_sp <-
  balance_dataset(some_sp, response = "ind_ha", absence_ratio = 0.2)

# Fit a RAF model
mraf <- fit_abund_raf(
  data = some_sp,
  response = "ind_ha",
  predictors = c("bio12", "elevation", "sand"),
  predictors_f = c("eco"),
  partition = ".part",
  mtry = 3,
  ntree = 500,
  predict_part = TRUE
)

mraf
}
}
\seealso{
\code{\link{fit_abund_cnn}},
\code{\link{fit_abund_dnn}},
\code{\link{fit_abund_gam}},
\code{\link{fit_abund_gbm}},
\code{\link{fit_abund_glm}},
\code{\link{fit_abund_net}},
\code{\link{fit_abund_svm}},
\code{\link{fit_abund_xgb}}
}
