% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_abund_xgt.R
\name{fit_abund_xgt}
\alias{fit_abund_xgt}
\title{Fit and validate Extreme Gradient Boosting models}
\usage{
fit_abund_xgt(
  data,
  response,
  predictors,
  predictors_f,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  params,
  nrounds = 500,
  verbose = TRUE
)
}
\arguments{
\item{data}{}

\item{response}{}

\item{predictors}{}

\item{predictors_f}{}

\item{fit_formula}{}

\item{partition}{}

\item{predict_part}{}

\item{params}{}

\item{nrounds}{}

\item{verbose}{}
}
\description{
Fit and validate Extreme Gradient Boosting models
}
