% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_linked_tables.R
\name{flatten}
\alias{flatten}
\title{Flatten a data model}
\usage{
flatten(dm)
}
\arguments{
\item{dm}{A dm.}
}
\value{
A tbl.
}
\description{
Flatten a data model
}
\examples{
library(dm)

dm <- adm_demo()
dm |> flatten_linked_tables()

dm |>
  dm_select_tbl(a, b, a_b) |>
  flatten_linked_tables()
}
