% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_linked_tables.R
\name{flatten_linked_tables}
\alias{flatten_linked_tables}
\title{Flatten a data model}
\usage{
flatten_linked_tables(dm, .join = dplyr::left_join)
}
\arguments{
\item{dm}{A \code{\link[dm]{dm}} object.}

\item{.join}{The type of join to be performed, see \code{\link[dplyr:mutate-joins]{dplyr::join()}}.}
}
\value{
A single table that results from consecutively joining all linked
tables.
}
\description{
Flatten a data model
}
\examples{
library(dm)

dm <- adm_demo()
dm \%>\% flatten_linked_tables()

dm \%>\%
  dm_select_tbl(a, b, a_b) \%>\%
  flatten_linked_tables()
}
