% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_arch_list.R
\name{generate_arch_list}
\alias{generate_arch_list}
\title{Title}
\usage{
generate_arch_list(
  type,
  number_of_features,
  number_of_outputs,
  n_layers = c(1, 2),
  n_neurons = c(7),
  batch_norm = TRUE,
  dropout = 0.2,
  sample_size = c(11, 11),
  number_of_fc_layers = 1,
  fc_layers_size = c(14)
)
}
\arguments{
\item{fc_layers_size}{}
}
\description{
Title
}
