% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cnn_architecture.R
\name{generate_cnn_architecture}
\alias{generate_cnn_architecture}
\title{Generate Convolutional Neural Network architecture}
\usage{
generate_cnn_architecture(
  number_of_features = 7,
  number_of_outputs = 1,
  sample_size = c(11, 11),
  number_of_conv_layers = 2,
  conv_layers_size = c(14, 28),
  conv_layers_kernel = 3,
  conv_layers_stride = 1,
  conv_layers_padding = 0,
  number_of_fc_layers = 1,
  fc_layers_size = c(28),
  pooling = FALSE,
  batch_norm = TRUE,
  dropout = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Generate Convolutional Neural Network architecture
}
