% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dnn_architecture.R
\name{generate_dnn_architecture}
\alias{generate_dnn_architecture}
\title{Title}
\usage{
generate_dnn_architecture(
  number_of_features = 7,
  number_of_outputs = 1,
  number_of_hidden_layers = 2,
  hidden_layers_size = c(14, 7),
  verbose = FALSE
)
}
\arguments{
\item{number_of_features}{}

\item{number_of_outputs}{}

\item{number_of_hidden_layers}{}

\item{hidden_layers_size}{}

\item{verbose}{}
}
\description{
Title
}
