% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_abund_bpdp.R
\name{p_abund_bpdp}
\alias{p_abund_bpdp}
\title{ADM Bivariate partial dependence plot}
\usage{
p_abund_bpdp(
  model,
  predictors = NULL,
  resolution = 50,
  training_data = NULL,
  training_boundaries = NULL,
  projection_data = NULL,
  clamping = FALSE,
  color_gradient = c("#000004", "#1B0A40", "#4A0C69", "#781B6C", "#A42C5F", "#CD4345",
    "#EC6824", "#FA990B", "#F7CF3D", "#FCFFA4"),
  color_training_boundaries = "white",
  theme = ggplot2::theme_classic()
)
}
\arguments{
\item{theme}{}
}
\description{
ADM Bivariate partial dependence plot
}
