% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotfunctions.R
\name{process_and_visualize_meta_methods}
\alias{process_and_visualize_meta_methods}
\title{Process and Visualize Meta-Methods}
\usage{
process_and_visualize_meta_methods(mev.out, info, color_list, seed = 2024)
}
\arguments{
\item{mev.out}{Output from MEV method (assumed to contain $diffu.dist).}

\item{info}{A vector of true class labels.}

\item{color_list}{A list of colors for visualization.}

\item{seed}{An integer seed for reproducibility. Default is 2024.}
}
\value{
A list containing:
  \item{ARI_list}{A list with ARI and NMI scores for the ADM method.}
  \item{ASW_list}{A list with silhouette widths for the ADM method.}
  \item{umap_viz}{A ggplot object visualizing the UMAP projection.}
  \item{umap_adm}{UMAP coordinates for the ADM method.}
}
\description{
Process and Visualize Meta-Methods
}
\details{
This function processes and visualizes results from the ADM (Adaptive Distance Matrix) meta-method.
It performs UMAP dimensionality reduction, calculates ARI (Adjusted Rand Index) and NMI (Normalized Mutual Information) scores,
and computes silhouette widths. The number of clusters (k) is automatically determined from the unique values in the info vector.
}
\note{
This function requires the following packages: uwot, cluster, ggplot2
It also relies on two custom functions: cal_ari_nmi() and visualization_func()
}
\examples{
# This function requires specific input structures and external functions
# An example cannot be easily provided without those dependencies

}
