% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_arch_list.R
\name{select_arch_list}
\alias{select_arch_list}
\title{Select arch list}
\usage{
select_arch_list(
  arch_list,
  type = c("dnn", "cnn"),
  method = "percentile",
  n_samples = 1,
  min_max = TRUE
)
}
\arguments{
\item{arch_list}{}

\item{type}{}

\item{method}{}

\item{n_samples}{}

\item{min_max}{}
}
\description{
Select arch list
}
