% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_abund_cnn_test.R
\name{tune_abund_cnn_test}
\alias{tune_abund_cnn_test}
\title{TEST VERSION Fit and validate Convolutional Neural Network with exploration of hyper-parameters that optimize performance}
\usage{
tune_abund_cnn_test(
  data,
  response,
  predictors,
  predictors_f = NULL,
  x,
  y,
  rasters,
  crop_size = 5,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  grid = NULL,
  architectures = NULL,
  metrics = NULL,
  n_cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{x}{character. Column name with longitude data}

\item{y}{character. Column name with latitude data}

\item{rasters}{SpatRaster. Raster with TODO}

\item{verbose}{}
}
\description{
TEST VERSION Fit and validate Convolutional Neural Network with exploration of hyper-parameters that optimize performance
}
