% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_abund_dnn.R
\name{tune_abund_dnn}
\alias{tune_abund_dnn}
\title{Fit and validate Deep Neural Network model with exploration of hyper-parameters that optimize performance}
\usage{
tune_abund_dnn(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  grid = NULL,
  architectures = NULL,
  metrics = NULL,
  n_cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Fit and validate Deep Neural Network model with exploration of hyper-parameters that optimize performance
}
