% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_abund_glm.R
\name{tune_abund_glm}
\alias{tune_abund_glm}
\title{Fit and validate Generalized Linear Models with exploration of hyper-parameters that optimize performance}
\usage{
tune_abund_glm(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  sigma_formula = ~1,
  nu_formula = ~1,
  tau_formula = ~1,
  partition,
  predict_part = FALSE,
  grid = NULL,
  metrics = NULL,
  n_cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{tibble or data.frame. Database with response, predictors, and partition values}

\item{response}{character. Column name with species abundance.}

\item{predictors}{character. Vector with the column names of quantitative predictor variables (i.e. continuous variables). Usage predictors = c("temp", "precipt", "sand")}

\item{predictors_f}{character. Vector with the column names of qualitative predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")}

\item{fit_formula}{formula. A formula object with response and predictor variables (e.g. formula(abund ~ temp + precipt + sand + landform)). Note that the variables used here must be consistent with those used in response, predictors, and predictors_f arguments. Default NULL}

\item{partition}{character. Column name with training and validation partition groups.}

\item{predict_part}{logical. Save predicted abundance for testing data. Default = FALSE}

\item{grid}{tibble or data.frame. A dataframe with "distribution", "poly", "inter_order" as columns and its values combinations as rows.}

\item{metrics}{character. Vector with one or more metrics from c("corr_spear","corr_pear","mae","pdisp","inter","slope").}

\item{n_cores}{numeric. Number of cores used in parallel processing.}

\item{verbose}{logical. If FALSE, disables all console messages. Default TRUE}
}
\value{
A list object with:
\itemize{
\item model: A "gamlss" object from gamlss package. This object can be used to predicting.
\item predictors: A tibble with quantitative (c column names) and qualitative (f column names) variables use for modeling.
\item performance: A tibble with selected model's performance metrics calculated in adm_eval.
\item performance_part: A tibble with performance metrics for each test partition.
\item predicted_part: A tibble with predicted abundance for each test partition.
\item optimal_combination: A tibble with the selected hyperparameter combination and its performance.
\item all_combinations: A tibble with all hyperparameters combinations and its performance.
}
}
\description{
Fit and validate Generalized Linear Models with exploration of hyper-parameters that optimize performance
}
\examples{
\dontrun{
require(terra)
require(dplyr)
require(gamlss)

# Database with species abundance and x and y coordinates
data("sppabund")

# Extract data for a single species
some_sp <- sppabund \%>\%
  dplyr::filter(species == "Species one") \%>\%
  dplyr::select(-.part2, -.part3)

# Explore reponse variables
some_sp$ind_ha \%>\% range()
some_sp$ind_ha \%>\% hist()

# Here we balance number of absences
some_sp <-
  balance_dataset(some_sp, response = "ind_ha", absence_ratio = 0.2)

# Explore different family distributions
suitable_distributions <- family_selector(data = some_sp, response = "ind_ha")
suitable_distributions

# Create a grid
glm_grid <- expand.grid(
  poly = c(2, 3),
  inter_order = c(1, 2),
  distribution = suitable_distributions$family_call
)

# Tune a GLM model
tuned_glm <- tune_abund_glm(
  data = some_sp,
  response = "ind_ha",
  predictors = c("bio12", "elevation", "sand"),
  fit_formula = formula("ind_ha ~ bio12 + elevation + sand + eco"),
  sigma_formula = formula("ind_ha ~ bio12 + elevation"),
  nu_formula = formula("ind_ha ~ bio12 + elevation"),
  predictors_f = c("eco"),
  partition = ".part",
  predict_part = TRUE,
  metrics = c("corr_pear", "mae"),
  grid = glm_grid,
  n_cores = 3
)

tuned_glm
}
}
