% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_abund_raf.R
\name{tune_abund_raf}
\alias{tune_abund_raf}
\title{Fit and validate Random Forest models with exploration of hyper-parameters that optimize performance}
\usage{
tune_abund_raf(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  grid = NULL,
  metrics = NULL,
  n_cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{data.frame. Database with response (i.e., abundance) and predictors values.}

\item{response}{character. Column name with species abundance data (e.g., 0, 1, 45).}

\item{predictors}{character. Vector with the column names of quantitative predictor variables
(i.e. continuous variables). Usage predictors = c("temperature", "sand", "elevation")}

\item{predictors_f}{character. Vector with the column names of qualitative predictor variables
(i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")}

\item{fit_formula}{formula. A formula object with response and predictor
variables (e.g. formula(abund ~ temp + sand + pH + landform)).
Note that the variables used here must be consistent with those used in
response, predictors, and predictors_f arguments. Default NULL}

\item{partition}{character. Column name with training and validation partition groups.}

\item{predict_part}{logical. Save predicted abundance for testing data. Default = FALSE}

\item{grid}{data.frame. A data frame object with algorithm hyper-parameters values to be tested.
It is recommended to generate this data.frame with the grid() function. Hyper-parameter needed
for tuning is 'mtry' and 'ntree'. The maximum mtry cannot exceed the total number of predictors.}

\item{metrics}{character. Performance metric used for selecting the best
combination of hyper-parameter values. One or more of the following metrics can be
 used: xxx, xxx, xxx (related to accuracy), xxx, xxx, xxx (related to discrimination)
xxx, xxx, xxx (related to precision). Usage metrics = c(corr_spear, slope)}

\item{n_cores}{numeric. Number of cores use for parallelization. Default 1}

\item{verbose}{}
}
\description{
Fit and validate Random Forest models with exploration of hyper-parameters that optimize performance
}
\examples{
\dontrun{
TODO
require(dplyr)
}
}
