% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_abund_svm.R
\name{tune_abund_svm}
\alias{tune_abund_svm}
\title{Fit and validate Support Vector Machine models with exploration of hyper-parameters that optimize performance}
\usage{
tune_abund_svm(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  grid = NULL,
  metrics = NULL,
  n_cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{}

\item{response}{}

\item{predictors}{}

\item{predictors_f}{}

\item{fit_formula}{}

\item{partition}{}

\item{predict_part}{}

\item{grid}{}

\item{metrics}{}

\item{n_cores}{}

\item{verbose}{}
}
\description{
Fit and validate Support Vector Machine models with exploration of hyper-parameters that optimize performance
}
