% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_abund_xgt.R
\name{tune_abund_xgt}
\alias{tune_abund_xgt}
\title{Fit and validate Extreme Gradient Boosting models with exploration of hyper-parameters that optimize performance}
\usage{
tune_abund_xgt(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  grid = NULL,
  objective,
  metrics = NULL,
  n_cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{data}{}

\item{response}{}

\item{predictors}{}

\item{predictors_f}{}

\item{fit_formula}{}

\item{partition}{}

\item{predict_part}{}

\item{grid}{}

\item{objective}{}

\item{metrics}{}

\item{n_cores}{}

\item{verbose}{}
}
\description{
Fit and validate Extreme Gradient Boosting models with exploration of hyper-parameters that optimize performance
}
