% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotfunctions.R
\name{visualize_individual_methods}
\alias{visualize_individual_methods}
\title{Visualize Individual Methods}
\usage{
visualize_individual_methods(
  matrices,
  method_names,
  info = NULL,
  color_list = NULL
)
}
\arguments{
\item{matrices}{A list of matrices, each representing a different method's output.}

\item{method_names}{A character vector specifying the names of the visualization methods.}

\item{info}{An optional vector of group labels for each data point. If not provided, all points will be the same color.}

\item{color_list}{An optional named vector of colors for each group in 'info'. If not provided, a default color palette will be used.}
}
\value{
A list of ggplot objects representing the scatter plots.
}
\description{
This function creates scatter plots for multiple matrices using different methods.
}
\examples{
matrices <- list(matrix(rnorm(200), 100, 2), matrix(rnorm(200), 100, 2))
method_names <- c("Method1", "Method2")
# Example with default coloring (all points same color)
plots1 <- visualize_individual_methods(matrices, method_names)

# Example with group information
info <- sample(c("A", "B", "C"), 100, replace = TRUE)
color_list <- c("A" = "red", "B" = "blue", "C" = "green")
plots2 <- visualize_individual_methods(matrices, method_names, info, color_list)

}
