% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm.R
\name{adm}
\alias{adm}
\title{Create a data model with automatically generated primary and foreign keys}
\usage{
adm(dir)
}
\arguments{
\item{dir}{A directory with .csv tables.}
}
\value{
A "dm" object.
}
\description{
Create a data model with automatically generated primary and foreign keys
}
\examples{
library(tibble)
library(withr)
library(readr)
library(fs)

tmp <- local_tempdir()
write_csv(tibble(x_id = 1, y_id = 2), path(tmp, "x.csv"))
write_csv(tibble(y_id = 2), path(tmp, "y.csv"))

adm(tmp)
}
