% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_eval.R
\name{adm_eval}
\alias{adm_eval}
\title{Calculate different model performance metrics}
\usage{
adm_eval(obs, pred)
}
\arguments{
\item{obs}{numeric. Observed abundance}

\item{pred}{numeric. Predicted abundance}
}
\value{
a tibble with next columns: corr_spear, corr_pear, mae, inter, slope, and pdispersion (see details)
}
\description{
Calculate difernet model performance related to model accuracy, discrimination, and precision.
}
\details{
This function calculate metric related to:
\itemize{
  \item Accuracy: mean absolute error (mae)
  TODO explain inter = inter slope = slope,
  \item Discrimination: Spearman’s rank correlation (corr_spear) and Pearson’s correlation (corr_pear)
  \item Precision:  Ratio between SD of predicted and observed abundance (pdispersion),
  }
  Further deatails see  Waldock et al. (2022)
}
\examples{
\dontrun{
require(dplyr)
TODO
}
}
\references{
\itemize{
  \item Waldock, C., Stuart-Smith, R.D., Albouy, C., Cheung, W.W.L., Edgar, G.J., Mouillot, D., Tjiputra, J., Pellissier, L., 2022. A quantitative review of abundance-based species distribution models. Ecography https://doi.org/10.1111/ecog.05694 }
}
