% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_summarize.R
\name{adm_summarize}
\alias{adm_summarize}
\title{Merge model performance tables}
\usage{
adm_summarize(models)
}
\arguments{
\item{models}{list. A list a single or several models fitted with some of fit_ or tune_ functions. Usage models = list(mod1, mod2, mod3)}
}
\value{
Combined model performance table for all input models. Models fit with tune will include model performance for the best hyperparameters.
}
\description{
Merge model performance tables
}
\examples{
\dontrun{
require(flexsdm)
data("sppabund")

sp1 <- sppabund \%>\%
  dplyr::filter(species == "Species one", ind_ha > 0) \%>\%
  dplyr::mutate(ind_ha = ind_ha \%>\% round())
# Families for GAM and GLM
dis_f <- system.file("external/families_bank.txt", package = "adm") \%>\%
  utils::read.delim(., header = TRUE, quote = "\t") \%>\%
  dplyr::as_tibble()
dis_f$family_name

# Fit GAM
m_gam <- fit_abund_gam(
  data = sp1,
  response = "ind_ha",
  predictors = c("bio1", "bio12", "bio15", "bio3", "cfvo", "elevation"),
  partition = ".part",
  family = "PO",
  inter = 1
)

m_svm <- fit_abund_svm(
  data = sp1,
  response = "ind_ha",
  predictors = c("bio1", "bio12", "bio15", "bio3", "cfvo", "elevation"),
  partition = ".part"
)

m_raf <- fit_abund_raf(
  data = sp1,
  response = "ind_ha",
  predictors = c("bio1", "bio12", "bio15", "bio3", "cfvo", "elevation"),
  partition = ".part"
)

adm_summarize(list(m_gam, m_svm, m_raf))
}
}
