% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_uncertainty.R
\name{adm_uncertainty}
\alias{adm_uncertainty}
\title{Calculate abundance distribution model uncertainty}
\usage{
adm_uncertainty(
  models,
  training_data,
  response,
  pred,
  iteration = 50,
  n_cores = 1
)
}
\arguments{
\item{models}{A model object from `fit_abund_*` or `tune_abund_*` functions.}

\item{training_data}{A data.frame or tibble with abundance data and predictors.}

\item{response}{character. Column name of the response variable.}

\item{pred}{A SpatRaster object with the environmental layers for projection.}

\item{iteration}{numeric. The number of bootstrap iterations. Default 50.}

\item{n_cores}{numeric. The number of cores to use for parallel processing. Default 1.}

\item{...}{Additional arguments passed to refitting functions or \code{\link{adm_predict}}
(e.g., \code{x}, \code{y}, \code{rasters}, \code{sample_size} for CNN; \code{custom_architecture} for DNN/CNN;
\code{invert_transform}, \code{transform_negative} for spatial prediction).}
}
\value{
A SpatRaster object with a single layer representing the model uncertainty, calculated as the standard deviation of the bootstrap predictions.
}
\description{
This function calculates the uncertainty of an abundance distribution model by performing a bootstrap procedure.
It refits the model multiple times on resampled data and then calculates the standard deviation of the predictions across
all iterations.
}
\examples{
\dontrun{
require(terra)
require(dplyr)

# Load data
data("cretusa_data")
cretusa_predictors <- system.file("external/cretusa_predictors.tif", package = "adm")
cretusa_predictors <- terra::rast(cretusa_predictors)

species_data <- adm_extract(
  data = cretusa_data,
  x = "x",
  y = "y",
  env_layer = cretusa_predictors
)

# Fit model
mraf <- fit_abund_raf(
  data = species_data,
  response = "ind_ha",
  predictors = c("PC1", "PC2", "PC3"),
  partition = ".part"
)

# Calculate uncertainty
unc <- adm_uncertainty(
  models = mraf,
  training_data = species_data,
  response = "ind_ha",
  pred = cretusa_predictors[[c("PC1", "PC2", "PC3")]],
  iteration = 10,
  n_cores = 2
)

plot(unc)
}
}
\seealso{
\code{\link{adm_predict}}
}
