% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_dataset.R
\name{balance_dataset}
\alias{balance_dataset}
\title{Balance dataset at a given presence:absence ratio}
\usage{
balance_dataset(data, response, absence_ratio)
}
\arguments{
\item{data}{}

\item{response}{}

\item{absence_ratio}{}
}
\description{
Balance dataset at a given presence:absence ratio
}
