% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_make_samples.R
\name{cnn_make_samples}
\alias{cnn_make_samples}
\title{TODO cnn_make_samples}
\usage{
cnn_make_samples(df, x, y, response, raster, size = 5)
}
\arguments{
\item{size}{}
}
\description{
TODO cnn_make_samples
}
