% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnn_make_samples.R
\name{cnn_make_samples}
\alias{cnn_make_samples}
\title{cnn_make_samples}
\usage{
cnn_make_samples(df, longitude, latitude, response, raster, size = 5)
}
\arguments{
\item{df}{}

\item{longitude}{}

\item{latitude}{}

\item{response}{}

\item{raster}{}

\item{size}{}
}
\description{
cnn_make_samples
}
