% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_abund_bpdp.R
\name{data_abund_bpdp}
\alias{data_abund_bpdp}
\title{Calculate data to construct bivariate partial dependence plots}
\usage{
data_abund_bpdp(
  model,
  predictors,
  resolution = 50,
  training_data = NULL,
  invert_transform = NULL,
  response_name = NULL,
  training_boundaries = NULL,
  projection_data = NULL
)
}
\arguments{
\item{model}{object returned by any fit_abund or tune_abund family functions}

\item{predictors}{character. Vector with two predictor name(s) to plot. If NULL all predictors will be plotted. Default NULL}

\item{resolution}{numeric. Number of equally spaced points at which to predict continuous predictors. Default 50}

\item{training_data}{data.frame or tibble. Database with response (0,1) and predictor values used to fit a model. Default NULL}

\item{training_boundaries}{character. Plot training conditions boundaries based on training data (i.e., presences, presences and absences, etc). If training_boundaries = "convexh", function will delimit training environmental region based on a convex-hull. If training_boundaries = "rectangle", function will delimit training environmental region based on four straight lines. If used any methods it is necessary provide data in training_data argument. If NULL all predictors will be used. Default NULL.}

\item{projection_data}{SpatRaster. Raster layer with environmental variables used for model projection. Default NULL}
}
\value{
A list with two tibbles "pdpdata" and "resid".
\itemize{
\item pspdata: has data to construct partial dependence surface plot, the first two column includes
values of the selected environmental variables, the third column with predicted suitability.
\item training_boundaries: has data to plot boundaries of training data.
}
}
\description{
Calculate data to construct bivariate partial dependence for two predictor set
}
\seealso{
\code{\link{data_abund_pdp},  \link{p_abund_bpdp}, \link{p_abund_pdp}}
}
