% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_abund_pdp.R
\name{data_abund_pdp}
\alias{data_abund_pdp}
\title{Calculate data to construct partial dependence plots}
\usage{
data_abund_pdp(
  model,
  predictors,
  resolution = 50,
  resid = FALSE,
  training_data = NULL,
  projection_data = NULL
)
}
\arguments{
\item{model}{}

\item{predictors}{}

\item{resolution}{}

\item{resid}{}

\item{training_data}{}

\item{projection_data}{}
}
\description{
Calculate data to construct partial dependence plots
}
