% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{family_selector}
\alias{family_selector}
\title{Select family distribution for GAM and GLM}
\usage{
family_selector(data, response)
}
\arguments{
\item{response}{}
}
\description{
Select family distribution suited for a given response variables (e.g., count, zero-inflated) used to fit GAM and GLM models
}
