% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_abund_cnn.R
\name{fit_abund_cnn}
\alias{fit_abund_cnn}
\title{Fit and train Convolutional Neural Network Model}
\usage{
fit_abund_cnn(
  data,
  response,
  predictors,
  predictors_f = NULL,
  longitude,
  latitude,
  rasters,
  crop_size = 5,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  learning_rate = 0.01,
  n_epochs = 10,
  batch_size = 32
)
}
\arguments{
\item{data}{tibble or data.frame. Database with response, predictors, and partition values}

\item{response}{character. Column name with species abundance.}

\item{predictors}{character. Vector with the column names of quantitative predictor variables (i.e. continuous variables). Usage predictors = c("temp", "precipt", "sand")}

\item{predictors_f}{character. Vector with the column names of qualitative predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")}

\item{longitude}{character. The name of the column containing longitude information for each observation.}

\item{latitude}{character. The name of the column containing latitude information for each observation.}

\item{rasters}{a terra SpatRaster object. A raster containing the predictor variables to be cropped around each observation.}

\item{crop_size}{numeric. An integer defining the range of pixels around the observation from the raster object passed to rasters parameter. Default = 5}

\item{fit_formula}{formula. A formula object with response and predictor variables (e.g. formula(abund ~ temp + precipt + sand + landform)). Note that the variables used here must be consistent with those used in response, predictors, and predictors_f arguments. Default NULL}

\item{partition}{character. Column name with training and validation partition groups.}

\item{predict_part}{logical. Save predicted abundance for testing data. Default = FALSE}

\item{learning_rate}{numeric. The size of the step taken during the optimization process. Default = 0.01}

\item{n_epochs}{numeric. How many times the learning algorithm will work through the training set. Default = 10}

\item{batch_size}{numeric. A batch is a subset of the training set used in a single iteration of the training process. The size of each batch is referred to as the batch size. Default = 32}
}
\description{
Fit and train Convolutional Neural Network Model
}
