% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_abund_gam.R
\name{fit_abund_gam}
\alias{fit_abund_gam}
\title{Fit and validate Generalized Additive Models}
\usage{
fit_abund_gam(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  family,
  inter = "automatic"
)
}
\arguments{
\item{method}{}
}
\description{
Fit and validate Generalized Additive Models
}
