% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_abund_gam.R
\name{fit_abund_gam}
\alias{fit_abund_gam}
\title{fit_abund_gam}
\usage{
fit_abund_gam(
  data,
  response,
  predictors,
  predictors_f,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  k = 10,
  family = "poisson",
  method = "REML"
)
}
\arguments{
\item{data}{}

\item{response}{}

\item{predictors}{}

\item{predictors_f}{}

\item{fit_formula}{}

\item{partition}{}

\item{predict_part}{}

\item{k}{}

\item{family}{}

\item{method}{}
}
\description{
fit_abund_gam
}
