% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_abund_raf.R
\name{fit_abund_raf}
\alias{fit_abund_raf}
\title{Fit and validate Random Forests models}
\usage{
fit_abund_raf(
  data,
  response,
  predictors,
  predictors_f,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  mtry = sqrt(length(c(predictors, predictors_f))),
  ntree = 500
)
}
\arguments{
\item{data}{tibble or data.frame. Database with response, predictors, and partition values}

\item{response}{character. Column name with species abundance.}

\item{predictors}{character. Vector with the column names of quantitative
predictor variables (i.e. continuous variables). Usage predictors = c("temp", "precipt", "sand")}

\item{predictors_f}{character. Vector with the column names of qualitative predictor
variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")}

\item{fit_formula}{formula. A formula object with response and predictor variables (e.g. formula(abund ~ temp + precipt + sand + landform)). Note that the variables used here must be consistent with those used in response, predictors, and predictors_f arguments. Default NULL}

\item{partition}{character. Column name with training and validation partition groups.}

\item{predict_part}{logical. Save predicted abundance for testing data. Default = FALSE}

\item{mtry}{numeric. Number of variables randomly sampled as candidates at each split. Default sqrt(length(c(predictors, predictors_f)))}
}
\value{
A list object with:
\itemize{
\item model: A "randomForest" class object from randomForest package. This object can be used for predicting.
\item predictors: A tibble with quantitative (c column names) and qualitative (f column names) variables use for modeling.
\item performance: Averaged performance metrics (see \code{\link{adm_eval}}).
\item performance_part: Performance metrics for each partition.
\item predicted_part: Observed and predicted abundance for each test partition.
}
}
\description{
Fit and validate Random Forests models
}
