% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_abund_xgt.R
\name{fit_abund_xgt}
\alias{fit_abund_xgt}
\title{Fit and validate Extreme Gradient Boosting models}
\usage{
fit_abund_xgt(
  data,
  response,
  predictors,
  predictors_f,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  nrounds = 100,
  max_depth = 4,
  eta = 0.2,
  gamma = 5,
  colsample_bytree = 1,
  min_child_weight = 1,
  subsample = 0.5,
  objective,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Fit and validate Extreme Gradient Boosting models
}
