% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cnn_architecture.R
\name{generate_cnn_architecture}
\alias{generate_cnn_architecture}
\title{Generate Convolutional Neural Network architecture}
\usage{
generate_cnn_architecture(
  number_of_features = 7,
  number_of_outputs = 1,
  sample_size = c(11, 11),
  number_of_conv_layers = 2,
  conv_layers_size = c(14, 28),
  conv_layers_kernel = rep(3, number_of_conv_layers),
  conv_layers_stride = rep(1, number_of_conv_layers),
  conv_layers_padding = rep(0, number_of_conv_layers),
  number_of_fc_layers = 1,
  fc_layers_size = c(28),
  verbose = FALSE
)
}
\arguments{
\item{number_of_features}{}

\item{number_of_outputs}{}

\item{sample_size}{}

\item{number_of_conv_layers}{}

\item{conv_layers_size}{}

\item{conv_layers_kernel}{}

\item{conv_layers_stride}{}

\item{conv_layers_padding}{}

\item{number_of_fc_layers}{}

\item{fc_layers_size}{}

\item{verbose}{}
}
\description{
Generate Convolutional Neural Network architecture
}
