% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dnn_architecture.R
\name{generate_dnn_architecture}
\alias{generate_dnn_architecture}
\title{Generate architectures for Deep Neural Network}
\usage{
generate_dnn_architecture(
  number_of_features = 7,
  number_of_outputs = 1,
  number_of_hidden_layers = 2,
  hidden_layers_size = c(14, 7),
  batch_norm = TRUE,
  dropout = 0,
  verbose = FALSE
)
}
\arguments{
\item{number_of_features}{numeric. Value that specifies the number of features in the dataset.}

\item{number_of_outputs}{numeric. Value that specifies the number of outputs.}

\item{number_of_hidden_layers}{numeric. Number of hidden layers in the neural network. Default 2.}

\item{hidden_layers_size}{numeric vector. Size of each hidden layer in the neural network. Default c(14, 7).}

\item{batch_norm}{logical. Whether to include batch normalization layers. Default TRUE.}

\item{dropout}{logical. Specifies whether dropout is included in the architecture. Default FALSE.}

\item{verbose}{logical. Whether to print the architecture. Default FALSE.}
}
\value{
A list containing:
\itemize{
\item net: a instantiated torch neural net.
\item arch: a string with a R expression to instantiate the neural network.
\item arch_dict: a list with a matrix describing the architecture structure.
}
}
\description{
Generate architectures for Deep Neural Network
}
\seealso{
\code{\link{generate_arch_list}}, \code{\link{generate_cnn_architecture}}
}
