% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{get_partition_samples}
\alias{get_partition_samples}
\title{Construct CNN samples list to use with tune_abund_cnn and fit_abund_cnn}
\usage{
get_partition_samples(
  data,
  x,
  y,
  response,
  folds,
  partition,
  rasters,
  crop_size
)
}
\value{
a list of arrays
}
\description{
Construct CNN samples list to use with tune_abund_cnn and fit_abund_cnn
}
