% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{model_selection}
\alias{model_selection}
\title{Title TODO}
\usage{
model_selection(hyper_combinations, metrics)
}
\arguments{
\item{hyper_combinations}{}

\item{metrics}{}
}
\description{
Title TODO
}
