% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_abund_pdp.R
\name{p_abund_pdp}
\alias{p_abund_pdp}
\title{ADM Partial Dependent Plot}
\usage{
p_abund_pdp(
  model,
  predictors = NULL,
  resolution = 100,
  resid = FALSE,
  training_data = NULL,
  projection_data = NULL,
  clamping = FALSE,
  rug = FALSE,
  colorl = c("#462777", "#6DCC57"),
  colorp = "black",
  alpha = 0.2,
  theme = ggplot2::theme_classic()
)
}
\arguments{
\item{theme}{}
}
\description{
ADM Partial Dependent Plot
}
