% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotfunctions.R
\name{process_and_visualize_meta_methods}
\alias{process_and_visualize_meta_methods}
\title{Process and Visualize Meta-Methods}
\usage{
process_and_visualize_meta_methods(
  mev.out,
  ensemble.out = NULL,
  info,
  k,
  color_list,
  seed = 2024
)
}
\arguments{
\item{mev.out}{Output from MEV method (assumed to contain $diffu.dist).}

\item{ensemble.out}{Output from an ensemble method (assumed to contain $ensemble.dist.mat). Can be NULL.}

\item{info}{A vector of true class labels.}

\item{k}{The number of clusters for k-means clustering.}

\item{color_list}{A named vector of colors for visualization, names should match unique values in info.}

\item{seed}{An integer seed for reproducibility. Default is 2024.}
}
\value{
A list containing:
  \item{ARI_list}{A list of data frames with ARI and NMI scores for each method}
  \item{ASW_list}{A list of average silhouette widths for each method}
  \item{umap_adm}{UMAP coordinates for ADM method}
  \item{plot}{A list of ggplot objects for each method}
  \item{umap_viz}{UMAP coordinates for meta-spec method (if ensemble.out is provided)}
}
\description{
Process and Visualize Meta-Methods
}
\details{
This function processes and visualizes results from two meta-methods: meta-spec and ADM.
If ensemble.out is NULL, only ADM method is executed.
It performs UMAP dimensionality reduction, k-means clustering, and calculates various metrics.
}
\note{
This function requires the following packages: uwot, mclust, aricode, cluster
}
\examples{
# This function requires specific input structures and external functions
# An example cannot be easily provided without those dependencies

}
