% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotfunctions.R
\name{process_and_visualize_meta_methods}
\alias{process_and_visualize_meta_methods}
\title{Process and Visualize Meta-Methods}
\usage{
process_and_visualize_meta_methods(
  ensemble.out,
  mev.out,
  info,
  k,
  color_list,
  seed = 2024
)
}
\arguments{
\item{ensemble.out}{Output from an ensemble method (assumed to contain $ensemble.dist.mat).}

\item{mev.out}{Output from MEV method (assumed to contain $diffu.dist).}

\item{info}{A vector of true class labels.}

\item{k}{The number of clusters for k-means clustering.}

\item{color_list}{A list of colors for visualization.}

\item{seed}{An integer seed for reproducibility. Default is 2024.}
}
\value{
A list containing ARI and NMI scores, silhouette widths, and UMAP coordinates for meta-spec and ADM methods.
}
\description{
Process and Visualize Meta-Methods
}
\details{
This function processes and visualizes results from two meta-methods: meta-spec and ADM.
It performs UMAP dimensionality reduction, k-means clustering, and calculates various metrics.
}
\note{
This function requires the following packages: umap, mclust, aricode, cluster
}
\examples{
# This function requires specific input structures and external functions
# An example cannot be easily provided without those dependencies

}
