% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{res_calculate}
\alias{res_calculate}
\title{Select probability distributions for GAM and GLM}
\usage{
res_calculate(
  type = c("layer", "pooling"),
  in_res,
  kernel_size,
  stride,
  padding
)
}
\arguments{
\item{kernel_size}{description}

\item{stride}{description}

\item{padding}{description}
}
\description{
Select probability distributions for GAM and GLM
}
