% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_abund_cnn.R
\name{tune_abund_cnn}
\alias{tune_abund_cnn}
\title{Fit and validate Convolutional Neural Network with exploration of hyper-parameters that optimize performance}
\usage{
tune_abund_cnn(
  data,
  response,
  predictors,
  predictors_f = NULL,
  x,
  y,
  rasters,
  sample_size,
  partition,
  predict_part = FALSE,
  grid = NULL,
  architectures = NULL,
  metrics = NULL,
  n_cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{data}{tibble or data.frame. Database with response, predictors, and partition values}

\item{response}{character. Column name with species abundance.}

\item{predictors}{character. Vector with the column names of quantitative predictor variables (i.e. continuous variables). Usage predictors = c("temp", "precipt", "sand")}

\item{predictors_f}{character. Vector with the column names of qualitative predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")}

\item{x}{character. Column name with longitude data.}

\item{y}{character. Column name with latitude data.}

\item{rasters}{character. Path to the raster file of environmental variables.}

\item{sample_size}{numeric. The dimension, in pixels, of raster samples. See cnn_make_samples beforehand. Default c(11,11)}

\item{partition}{character. Column name with training and validation partition groups.}

\item{predict_part}{logical. Save predicted abundance for testing data. Default = FALSE}

\item{grid}{tibble or data.frame. A dataframe with "batch_size", "n_epochs", "learning_rate" as columns and its values combinations as rows. If no grid is provided, function will create a default grid combining the next hyperparameters:
batch_size = 2^seq(4, 6)
n_epochs = 10
learning_rate = seq(from = 0.1, to = 0.2, by = 0.1)
validation_patience = 2
fitting_patience = 5. In case one or more hyperparameters are provided, the function will complete the grid with the default values.}

\item{architectures}{list or character. A list object containing a list of architectures (nn_modules_generators from torch), called "arch_list", and a list of matrices describing each architecture, called ("arch_dict"); use generate_arch_list function to create it. It's also possible to use "fit_intern", what will construct the default neural network architecture of fit_abund_cnn. If NULL, a list of architectures will be generated. Default NULL}

\item{metrics}{character. Vector with one or more metrics from c("corr_spear","corr_pear","mae","pdisp","inter","slope").}

\item{n_cores}{numeric. Number of cores used in parallel processing.}

\item{verbose}{logical. If FALSE, disables all console messages. Default TRUE}
}
\value{
A list object with:
\itemize{
\item model: A "luz_module_fitted" object from luz (torch framework). This object can be used to predicting.
\item predictors: A tibble with quantitative (c column names) and qualitative (f column names) variables use for modeling.
\item performance: A tibble with selected model's performance metrics calculated in adm_eval.
\item performance_part: A tibble with performance metrics for each test partition.
\item predicted_part: A tibble with predicted abundance for each test partition.
\item optimal_combination: A tibble with the selected hyperparameter combination and its performance.
\item all_combinations: A tibble with all hyperparameters combinations and its performance.
\item selected_arch: A numeric vector describing the selected architecture layers.
}
}
\description{
Fit and validate Convolutional Neural Network with exploration of hyper-parameters that optimize performance
}
