% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_abund_gam.R
\name{tune_abund_gam}
\alias{tune_abund_gam}
\title{Fit and validate Generalized Additive Models with exploration of hyper-parameters that optimize performance}
\usage{
tune_abund_gam(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  grid = NULL,
  metrics = NULL,
  n_cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{}

\item{response}{character. Column name with species abundance data (e.g., 0, 1, 45).}

\item{predictors}{character. Vector with the column names of quantitative predictor variables
(i.e. continuous variables). Usage predictors = c("temperature", "sand", "elevation")}

\item{predictors_f}{character. Vector with the column names of qualitative predictor variables
(i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")}

\item{fit_formula}{formula. A formula object with response and predictor
variables (e.g. formula(abund ~ temp + sand + pH + landform)).
Note that the variables used here must be consistent with those used in
response, predictors, and predictors_f arguments. Default NULL}

\item{partition}{character. Column name with training and validation partition groups.}

\item{predict_part}{}

\item{grid}{}

\item{metrics}{}

\item{n_cores}{}

\item{verbose}{}
}
\description{
Fit and validate Generalized Additive Models with exploration of hyper-parameters that optimize performance
}
