% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_abund_gbm.R
\name{tune_abund_gbm}
\alias{tune_abund_gbm}
\title{Title}
\usage{
tune_abund_gbm(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  grid = NULL,
  distribution,
  metrics = NULL,
  n_cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{}

\item{response}{}

\item{predictors}{}

\item{predictors_f}{}

\item{fit_formula}{}

\item{partition}{}

\item{predict_part}{}

\item{grid}{}

\item{distribution}{}

\item{metrics}{}

\item{n_cores}{}

\item{verbose}{}
}
\description{
Title
}
