% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_abund_gbm.R
\name{tune_abund_gbm}
\alias{tune_abund_gbm}
\title{Fit and validate Generalized Boosted Regression models with exploration of hyper-parameters that optimize performance}
\usage{
tune_abund_gbm(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition,
  predict_part = FALSE,
  grid = NULL,
  distribution,
  metrics = NULL,
  n_cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{data}{tibble or data.frame. Database with response, predictors, and partition values}

\item{response}{character. Column name with species abundance.}

\item{predictors}{character. Vector with the column names of quantitative predictor variables (i.e. continuous variables). Usage predictors = c("temp", "precipt", "sand")}

\item{predictors_f}{character. Vector with the column names of qualitative predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")}

\item{fit_formula}{formula. A formula object with response and predictor variables (e.g. formula(abund ~ temp + precipt + sand + landform)). Note that the variables used here must be consistent with those used in response, predictors, and predictors_f arguments. Default NULL}

\item{partition}{character. Column name with training and validation partition groups.}

\item{predict_part}{logical. Save predicted abundance for testing data. Default = FALSE}

\item{grid}{tibble or data.frame. A dataframe with "n.trees", "interaction.depth", "n.minobsinnode" and "shrinkage" as columns and its values as rows. If no grid is provided, function will create a default grid combining the next hyperparameters:
n.trees = c(100, 200, 300),
interaction.depth = c(1, 2, 3),
n.minobsinnode = c(5, 10, 15),
shrinkage = seq(0.001, 0.1, by = 0.05). In case one or more hyperparameters are provided, the function will complete the grid with the default values.}

\item{distribution}{character. A string specifying the distribution to be used. See gbm::gbm documentation for details.}

\item{metrics}{character. Vector with one or more metrics from c("corr_spear","corr_pear","mae","pdisp","inter","slope").}

\item{n_cores}{numeric. Number of cores used in parallel processing.}

\item{verbose}{logical. If FALSE, disables all console messages. Default TRUE}
}
\value{
A list object with:
\itemize{
\item model: A "gbm" object from gbm package. This object can be used to predicting.
\item predictors: A tibble with quantitative (c column names) and qualitative (f column names) variables use for modeling.
\item performance: A tibble with selected model's performance metrics calculated in adm_eval.
\item performance_part: A tibble with performance metrics for each test partition.
\item predicted_part: A tibble with predicted abundance for each test partition.
\item optimal_combination: A tibble with the selected hyperparameter combination and its performance.
\item all_combinations: A tibble with all hyperparameters combinations and its performance.
\item selected_arch: A numeric vector describing the selected architecture layers.
}
}
\description{
Fit and validate Generalized Boosted Regression models with exploration of hyper-parameters that optimize performance
}
\examples{
\dontrun{
require(dplyr)

data("sppabund")

some_sp <- sppabund \%>\%
  filter(species == "Species two")

gbm_grid <- expand.grid(
  interaction.depth = c(2, 4, 8, 16),
  n.trees = c(100, 500, 1000),
  n.minobsinnode = c(2, 5, 8),
  shrinkage = c(0.1, 0.5, 0.7)
)

tuned_gbm <- tune_abund_gbm(
  data = some_sp,
  response = "ind_ha",
  predictors = c("bio12", "elevation", "sand"),
  predictors_f = c("eco"),
  partition = ".part",
  predict_part = TRUE,
  metrics = c("corr_pear", "mae"),
  grid = gbm_grid,
  distribution = "gaussian",
  n_cores = 3
)

tuned_gbm$model
tuned_gbm$performance
tuned_gbm$optimal_combination
tuned_gbm$all_combinations
}
}
